french-ai-accounting-saas-backend-tmp/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Object Storage
boto3==1.29.7
google-cloud-storage==2.14.0
azure-storage-blob==12.19.0

# Message Queue
pika==1.3.2
kafka-python==2.0.2

# OCR Providers (optional for auth/login; Pillow build can fail on Windows without C++ build tools)
pytesseract==0.3.10
# Pillow==10.1.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
httpx==0.25.2

# Machine Learning & Anomaly Detection (anomaly service + run_monitoring_job.py)
scikit-learn>=1.3.0
numpy>=1.23,<2.0
pandas>=2.0.0

# Vector Database & Embeddings
pgvector==0.2.4
sentence-transformers>=2.6.1

# ERP Integration & SFTP
paramiko==3.4.0
pysftp==0.2.9

# SEPA (choose one)
sepaxml==2.7.0
# OR: lxml==5.1.0

# Performance & Caching
redis==5.0.1
aiohttp==3.9.1
